pyproject.toml
[project]
name = "notion-scripts"
version = "0.2"
description = "Sync and utility scripts for Notion and other services."
authors = [
    { name = "Andrei Hajdukewycz", email = "sancus@off.net" },
    { name = "Malini Das", email = "malini@thunderbird.net" },
    { name = "Philipp Kewisch", email = "kewisch@thunderbird.net" }
]
dependencies = [
    "notion-client>=2.2.1",
    "requests>=2.32.3",
    "sgqlc>=16.4",
    "sgqlc-schemas>=0.1.0",
    "python-dateutil>=2.9.0.post0",
    "md2notionpage==0.1.5",
    "notion-to-md-py==0.1.0",
]
readme = "README.md"
requires-python = ">= 3.12"

[tool.rye]
managed = true
virtual = true
dev-dependencies = []

[tool.rye.scripts]
sync = "python notion_sync.py"

[tool.ruff]
line-length = 120
target-version = "py312"
extend-exclude = ["libs/github_schema.py"]

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = []
extend-select = ["D"]
ignore = ["D100", "D203", "D213"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
exclude = ["libs/github_schema.py"]
